style: squish sorting modal to match Figma and move to filters directory
