+++ client/src/components/filters/ProgramFilter.jsx
@@ -17,7 +17,7 @@ import "./Sorting.css";
 
 const ProgramSortingModal = ({ onSortChange }) => {
   return (
-    <Menu>
+    <Menu closeOnSelect={true}>
       <MenuButton
         as={Button}
         variant="ghost"
@@ -35,21 +35,29 @@ const ProgramSortingModal = ({ onSortChange }) => {
         />
       </MenuButton>
       <MenuList className="program-sorting-popup">
-        <MenuItem onClick={() => onSortChange("title", "asc")}>
+        <MenuItem
+          onClick={() => onSortChange("title", "asc")}
+          width="100%"
+        >
           <Flex
             align="center"
             gap={2}
             className="sort-menu-item"
+            width="100%"
           >
             <TriangleUpIcon className="triangle-icon icon-style" />
             <Box className="icon-style">A‑Z</Box>
           </Flex>
         </MenuItem>
-        <MenuItem onClick={() => onSortChange("title", "desc")}>
+        <MenuItem
+          onClick={() => onSortChange("title", "desc")}
+          width="100%"
+        >
           <Flex
             align="center"
             gap={2}
             className="sort-menu-item"
+            width="100%"
           >
             <TriangleDownIcon className="triangle-icon icon-style" />
             <Box className="icon-style">Z‑A</Box>
